make auth by sanctum on backend

+++ backend/app/Http/Controllers/PostController.php
@@ -2,11 +2,22 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StorePostRequest;
+use App\Http\Requests\UpdatePostRequest;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -21,11 +32,11 @@ class PostController extends Controller
     public function store(Request $request)
     {
         $fields = $request->validate([
-            'title' => 'required:max:255',
+            'title' => 'required|max:255',
             'body' => 'required'
         ]);
 
-        $post = Post::create($fields);
+        $post = $request->user()->posts()->create($fields);
 
         return ['post' => $post];
     }
@@ -41,10 +52,12 @@ class PostController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Post $post)
+    public function update(UpdatePostRequest $request, Post $post)
     {
+        Gate::authorize('modify', $post);
+
         $fields = $request->validate([
-            'title' => 'required:max:255',
+            'title' => 'required|max:255',
             'body' => 'required'
         ]);
 
@@ -58,8 +71,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post);
+
         $post->delete();
 
-        return 'Post deleted';
+        return 'The post was deleted';
     }
 }
